+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Show creator for created cards on play (#49)
 - HSReplay: Add warning if tag is missing required entity id
 - Add Quest frame art
+- Support Shifting Blade
 
 ### Fixed
 - Fix fullscreen callback in Launcher not being called